initialized tables in the database

+++ models/blix_healthcare_db/silver/practitioner_taxonomy_code.py
@@ -26,7 +26,7 @@ class PractitionerTaxonomyCode(Base):
     __table_args__ = {"schema": "silver"}
 
     practitioner_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("silver.practitioner.practitioner_id"), primary_key=True, nullable=False)
-    taxonomy_code_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("silver.taxonomy_code.taxonomy_code_id"), primary_key=True, nullable=False)
+    taxonomy_code_id: Mapped[int] = mapped_column(Integer, ForeignKey("silver.taxonomy_code.taxonomy_code_id"), primary_key=True, nullable=False)
     preference_id: Mapped[Optional[int]] = mapped_column(Integer, ForeignKey("silver.preference.preference_id"), nullable=True)
     period_start_date: Mapped[date] = mapped_column(Date, nullable=False)
     period_end_date: Mapped[Optional[date]] = mapped_column(Date, nullable=True)
